feat: import prism dynamically / refactor
