+++ src/components/apiPage/styles.scss
@@ -87,37 +87,6 @@
   color: var(--white);
 }
 
-.api-block-example {
-  border: none;
-  border-bottom-left-radius: 3px;
-  border-bottom-right-radius: 3px;
-  border-top-left-radius: 0px;
-  border-top-right-radius: 0px;
-  margin-top: 0;
-  margin-bottom: 0;
-  font-size: 0.8rem;
-  padding: 0.75rem;
-}
-
-.api-block-example.request {
-  color: var(--white);
-  background-color: #2d2d2d;
-  border-radius: 4px;
-}
-
-.api-block-example.response {
-  background: #2d2d2d;
-  color: var(--white);
-}
-
-.api-params dd {
-  padding: 0;
-
-  p {
-    margin: 0;
-  }
-}
-
 .api-params dt {
   display: flex;
   flex-direction: row;
